% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dezim_pal.R
\name{dezim_pal}
\alias{dezim_pal}
\title{DeZIM Farbpaletten interpolieren}
\usage{
dezim_pal(palette = "alle", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name der Farbpalette (String / character). Zur Zeit "alle", "gruen", "drei"}

\item{reverse}{Boolean: Palette umdrehen?}

\item{...}{Zusätzliche Argumente, die an grDevices::colorRampPalette() weitergegeben werden}
}
\description{
DeZIM Farbpaletten interpolieren
}
\examples{
dezim_pal("alle")  # Liefert eine Funktion!
dezim_pal("alle")(8)
dezim_pal("alle")(8) |> colorspace::swatchplot()
dezim_pal("drei")(8) |> colorspace::swatchplot()
dezim_pal("gruen")(4) |> colorspace::swatchplot()
}
