% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dezim_cols.R
\name{dezim_cols}
\alias{dezim_cols}
\title{DeZIM Farben als Hexcodes extrahieren}
\usage{
dezim_cols(...)
}
\arguments{
\item{...}{Farbbezeichnung(en) als Strings (character)}
}
\description{
DeZIM Farben als Hexcodes extrahieren
}
\examples{
dezim_cols()
dezim_cols("Earth yellow")
dezim_cols("Earth yellow", "Dartmouth green")
ggplot(mtcars, aes(x = hp, y = mpg)) +
geom_point(size = 2.5, col = dezim_cols("Dartmouth green"))
}
